Remove the scatterPlot interface.
Renaming will come in the next CL (to remove "WebGL" from the filenames).
Change: 134457335
