+++ tensorflow/tensorboard/components/vz-projector/scatterPlotWebGLVisualizer.ts
@@ -14,7 +14,7 @@ limitations under the License.
 ==============================================================================*/
 
 import {RenderContext} from './renderContext';
-import {DataSet} from './scatterPlot';
+import {DataSet} from './scatterPlotWebGL';
 
 
 /**
