+++ tensorflow/tensorboard/components/vz-projector/scatterPlotWebGLVisualizerAxes.ts
@@ -14,7 +14,7 @@ limitations under the License.
 ==============================================================================*/
 
 import {RenderContext} from './renderContext';
-import {DataSet} from './scatterPlot';
+import {DataSet} from './scatterPlotWebGL';
 import {ScatterPlotWebGLVisualizer} from './scatterPlotWebGLVisualizer';
 
 const AXIS_COLOR = 0xb3b3b3;
